+++ dune/grid/io/file/gmshreader.hh
@@ -314,6 +314,7 @@ namespace Dune
     // exported data
     std::vector<int> boundary_id_to_physical_entity;
     std::vector<int> element_index_to_physical_entity;
+    std::vector<std::string> physical_entity_names;
 
     // static data
     static const int dim = GridType::dimension;
@@ -362,9 +363,16 @@ namespace Dune
       return boundary_id_to_physical_entity;
     }
 
+    //! Returns a map for the gmsh physical entity id (1-based index) for each entity of codim 0
     std::vector<int> & elementIndexMap()
     {
       return element_index_to_physical_entity;
+    }
+
+    //! Returns the names of the gmsh physical entities (0-based index)
+    std::vector<std::string> & physicalEntityNames()
+    {
+      return physical_entity_names;
     }
 
     void read (const std::string& f)
@@ -401,10 +409,32 @@ namespace Dune
       if (strcmp(buf,"$EndMeshFormat")!=0)
         DUNE_THROW(Dune::IOError, "expected $EndMeshFormat");
 
+      // physical name section
+      physical_entity_names.clear();
+      readfile(file,1,"%s\n",buf);
+      if (strcmp(buf,"$PhysicalNames")==0) {
+        int number_of_names;
+        readfile(file,1,"%d\n",&number_of_names);
+        if (verbose) std::cout << "file contains " << number_of_names << " physical entities" << std::endl;
+        physical_entity_names.resize(number_of_names);
+        std::string buf_name;
+        for( int i = 0; i < number_of_names; ++i ) {
+          int dim, id;
+          readfile(file,3, "%d %d %s\n", &dim, &id, buf);
+          buf_name.assign(buf);
+          auto begin = buf_name.find_first_of('\"') + 1;
+          auto end = buf_name.find_last_of('\"') - begin;
+          physical_entity_names[id-1].assign(buf_name.substr(begin, end));
+        }
+        readfile(file,1,"%s\n",buf);
+        if (strcmp(buf,"$EndPhysicalNames")!=0)
+          DUNE_THROW(Dune::IOError, "expected $EndPhysicalNames");
+        readfile(file,1,"%s\n",buf);
+      }
+
       // node section
       int number_of_nodes;
 
-      readfile(file,1,"%s\n",buf);
       if (strcmp(buf,"$Nodes")!=0)
         DUNE_THROW(Dune::IOError, "expected $Nodes");
       readfile(file,1,"%d\n",&number_of_nodes);
@@ -458,8 +488,8 @@ namespace Dune
         {
           int blub;
           readfile(file,1,"%d ",&blub);
-          // k == 1: physical entity (not used here)
-          // k == 2: elementary entity (not used here either)
+          // k == 1: physical entity
+          // k == 2: elementary entity (not used here)
           // if version_number < 2.2:
           //   k == 3: mesh partition 0
           // else